// src/app/core/main/delegations/interfaces/delegations.interface.ts
// Generated by https://quicktype.io

import { IClient } from "../../clients/interfaces/client.interface";
import { IUser } from "../../users/interfaces/user.interface";

export interface IDelegations {
    id: number;
    employeeId: number;
    consumerId: number;
}

export interface IDelegationsPost {
    employeeId: number;
    consumerId: number;
}

// Generated by https://quicktype.io

// Generated by https://quicktype.io

// Generated by https://quicktype.io

export interface IDelegationGet {
    id:         number;
    employeeId: number;
    consumerId: number;
    createdAt:  Date;
    updateAt:   Date;
    employee:   Employee;
    consumer:   Consumer;
}

export interface Consumer {
    id:         number;
    isCustomer: boolean;
    type:       string;
    person:     Person;
}

export interface Person {
    dni:       string;
    firstName: string;
    lastName:  string;
}

export interface Employee {
    id:       number;
    username: string;
    role:     string;
    person:   Person;
}